# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Education Planner API",
    "description": "\n    ## Динамический планировщик образовательной траектории\n    \n    API для студентов УрФУ, помогающее:\n    - 📚 Выбирать специальность и курсы\n    - 📊 Отслеживать развитие навыков\n    - 🎯 Получать персональные рекомендации\n    - 🔄 Моделировать смену специальности\n    \n    ### Основные возможности:\n    \n    - **Аутентификация**: JWT токены (access + refresh)\n    - **Специальности**: Просмотр roadmap и матрицы навыков\n    - **Курсы**: Запись, прохождение, отслеживание прогресса\n    - **Навыки**: Древовидная структура, gap-анализ\n    - **Рекомендации**: Умные подсказки по выбору курсов\n    - **What-If анализ**: Моделирование смены специальности\n    \n    ### Для начала работы:\n    1. Зарегистрируйтесь через `/api/auth/register`\n    2. Получите токен через `/api/auth/login`\n    3. Используйте токен в заголовке: `Authorization: Bearer \u003Ctoken\u003E`\n    ",
    "contact": {
      "name": "Education Planner Team",
      "email": "support@edu-planner.ru"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Регистрация нового пользователя",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Вход в систему",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Token",
        "description": "Обновление access токена с помощью refresh токена",
        "operationId": "refresh_token_api_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefresh"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Current User Info",
        "description": "Получение информации о текущем пользователе",
        "operationId": "get_current_user_info_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "description": "Смена пароля текущего пользователя.\n\nТребует:\n- Авторизацию (Bearer token)\n- Текущий пароль для подтверждения\n- Новый пароль (минимум 6 символов)\n- Подтверждение нового пароля",
        "operationId": "change_password_api_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/courses/": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get All Courses",
        "description": "Получить список курсов с фильтрами",
        "operationId": "get_all_courses_api_courses__get",
        "parameters": [
          {
            "name": "specialization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по специальности",
              "title": "Specialization Id"
            },
            "description": "Фильтр по специальности"
          },
          {
            "name": "semester",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по семестру",
              "title": "Semester"
            },
            "description": "Фильтр по семестру"
          },
          {
            "name": "is_elective",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по типу курса",
              "title": "Is Elective"
            },
            "description": "Фильтр по типу курса"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  },
                  "title": "Response Get All Courses Api Courses  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/my": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get My Courses",
        "description": "Получить курсы текущего студента",
        "operationId": "get_my_courses_api_courses_my_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudentCourseResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Courses Api Courses My Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/courses/{course_id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Detail",
        "description": "Получить детальную информацию о курсе",
        "operationId": "get_course_detail_api_courses__course_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/skills": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Skills",
        "description": "Получить навыки курса по уровням сложности",
        "operationId": "get_course_skills_api_courses__course_id__skills_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/enroll": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Enroll In Course",
        "description": "Записаться на курс",
        "operationId": "enroll_in_course_api_courses__course_id__enroll_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseEnrollRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentCourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/complete": {
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Complete Course",
        "description": "Завершить курс",
        "operationId": "complete_course_api_courses__course_id__complete_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentCourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/students/me/courses": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Courses",
        "description": "Получить курсы студента: завершенные, текущие и доступные",
        "operationId": "get_my_courses_api_students_me_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get My Courses Api Students Me Courses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/recommendations/courses": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Course Recommendations",
        "description": "Получить рекомендации курсов.\nНа первом этапе возвращает случайные курсы, восполняющие дефицит навыков.",
        "operationId": "get_course_recommendations_api_recommendations_courses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Количество рекомендаций",
              "default": 5,
              "title": "Limit"
            },
            "description": "Количество рекомендаций"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseDetailResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "credits": {
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "title": "Credits",
            "default": 3
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "is_elective": {
            "type": "boolean",
            "title": "Is Elective",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "difficulty_levels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Levels"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillInCourse"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CourseDetailResponse",
        "description": "Детальная информация о курсе"
      },
...
      "CourseResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "credits": {
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "title": "Credits",
            "default": 3
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "is_elective": {
            "type": "boolean",
            "title": "Is Elective",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "difficulty_levels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Levels"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CourseResponse",
        "description": "Схема ответа с курсом"
      },
      "CourseStatus": {
        "type": "string",
        "enum": [
          "enrolled",
          "in_progress",
          "completed",
          "dropped"
        ],
        "title": "CourseStatus",
        "description": "Статус прохождения курса"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChange": {
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6,
            "title": "New Password"
          },
          "confirm_password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6,
            "title": "Confirm Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password",
          "confirm_password"
        ],
        "title": "PasswordChange",
        "description": "Схема для смены пароля (требует текущий пароль)"
      },
      "PasswordChangeResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "message",
          "success"
        ],
        "title": "PasswordChangeResponse",
        "description": "Схема ответа после смены пароля"
      },
...
      "StudentCourseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "status": {
            "$ref": "#/components/schemas/CourseStatus"
          },
          "difficulty_level": {
            "type": "integer",
            "title": "Difficulty Level"
          },
          "grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "enrolled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Enrolled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "credits": {
            "type": "integer",
            "title": "Credits"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "course_name",
          "status",
          "difficulty_level",
          "enrolled_at",
          "credits"
        ],
        "title": "StudentCourseResponse",
        "description": "Информация о курсе студента"
      },
...
      "TokenRefresh": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "TokenRefresh",
        "description": "Схема для обновления токена"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse",
        "description": "Схема ответа с токенами"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "password"
        ],
        "title": "UserCreate",
        "description": "Схема для создания пользователя"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "description": "Схема для входа"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "minor_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Minor Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Схема ответа с данными пользователя"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}